06-Bluetooth/ruuvitag/requirements.txt
paho_mqtt
ruuvitag_sensor
